Move focus point to TwoFinger Gesture detector
