+++ library/src/main/java/com/almeros/android/multitouch/TwoFingerGestureDetector.java
@@ -2,7 +2,6 @@ package com.almeros.android.multitouch;
 
 import android.content.Context;
 import android.util.DisplayMetrics;
-import android.util.FloatMath;
 import android.view.MotionEvent;
 import android.view.ViewConfiguration;
 
@@ -36,10 +35,12 @@ public abstract class TwoFingerGestureDetector extends BaseGestureDetector {
 	protected float mPrevFingerDiffY;
 	protected float mCurrFingerDiffX;
 	protected float mCurrFingerDiffY;
-    
+    protected float mFocusX;
+    protected float mFocusY;
+
     private float mCurrLen;
     private float mPrevLen;
-	
+
     public TwoFingerGestureDetector(Context context) {
     	super(context);
     	
@@ -111,7 +112,33 @@ public abstract class TwoFingerGestureDetector extends BaseGestureDetector {
         }
         return mPrevLen;
     }
-    
+
+    /**
+     * Get gesture focus X point.
+     * @return focus X point of gesture.
+     */
+    public float getFocusX() {
+        return mFocusX;
+    }
+
+
+    /**
+     * Get gesture focus Y point.
+     * @return focus Y point of gesture.
+     */
+    public float getFocusY() {
+        return mFocusY;
+    }
+
+    /**
+     * Calculate focus point for gesture.
+     * @param curr current motion event used to calculate focus point.
+     */
+    protected void determineFocusPoint(MotionEvent curr) {
+        mFocusX = (curr.getX(0) + curr.getX(1)) * 0.5f;
+        mFocusY = (curr.getY(0) + curr.getY(1)) * 0.5f;
+    }
+
     /**
      * MotionEvent has no getRawX(int) method; simulate it pending future API approval. 
      * @param event
